**IMPORTANT: Please search among the [Pull request](../pulls) before creating one.**

## Title
Parallelize solving of the system when computing absorption probabilities.

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] New feature (non-breaking change which adds functionality)

## Description
<!--- Clearly and concisely describe your changes. -->
Added an option to solve the system in parallel, default is still sequential.

## How has this been tested?
No unit tests yet - I will add one test using parallel option.

##Benchmarking
For now, only on pancres, 20 lineages.
Before:
1.12 s ± 23.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
After:
954 ms ± 14.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

## Closes
closes #256 

@Marius1311 
There was previously this line: `x0 = None if ix in init_indices else x`, but it doesn't seem correct: https://github.com/theislab/cellrank/blob/master/cellrank/tools/estimators/_base_estimator.py#L419. Am I wrong?
```