@michalk8 2020-07-09
Also, you realize setting `._lin_probs` in this method overrides the computation done in `.{compute,set}_main_states()` in GPCCA?
Just checking whether it's intended.