@Marius1311 2020-07-09
Hi Mike, what's the issue with `x0 = None if ix in init_indices else x`? Basically, I want to say that for cells that belong to the same metastable state, use the previous solution as initialization because these problems are likely to have similar solutions. If however the last cell came from a different metastable state, don't use this initialization. 

You are right there may be an issue here: cells aren't ordered by metastable states in the matrix `B`, ie I'm not sure whether I'm really making use of this initialization at all!