@Marius1311 2020-07-09
The pancreas is too small though to benchmark this, can you try in lung or in the morris reprogramming data please? Because so far, there is no significant speed up.