@Marius1311 2020-07-09
I think the first thing to check here would really be: in the old formulation without parallelization, did we make use of clever initialization? Because if not, we should change that, and see how much improvement that gives. Once we have that, we can benchmark this against a parallelized version.