@Marius1311 2020-07-09
Okay, I think the `compute_absorption_probabilities` function needs a bit of work. Currently, it's too complicated. I will look into this hopefully later today, and refactor it. I will move the heavy lifting into a utility function we can test easily. Until then, I think it makes sense for you to pause your paralelization efforts.