@michalk8 2020-07-09
> Hi Mike, what's the issue with `x0 = None if ix in init_indices else x`? Basically, I want to say that for cells that belong to the same metastable state, use the previous solution as initialization because these problems are likely to have similar solutions.

Ah, I see, thanks for the explanation.

> Until then, I think it makes sense for you to pause your paralelization efforts.

Agreed.