@Marius1311 2020-07-09
I'm still on the refactoring, hopefully will be done tomorrow morning!